+++ tests/conftest.py
@@ -126,6 +126,26 @@ def configuration(smtpd):
                 "PASSWORD": smtpd.config.login_password,
                 "FROM_ADDR": "admin@mydomain.tld",
             },
+            "LOGGING": {
+                "version": 1,
+                "formatters": {
+                    "default": {
+                        "format": "[%(asctime)s] %(levelname)s in %(module)s: %(message)s",
+                    }
+                },
+                "handlers": {
+                    "wsgi": {
+                        "class": "logging.StreamHandler",
+                        "stream": "ext://sys.stdout",
+                        "formatter": "default",
+                    }
+                },
+                "root": {"level": "DEBUG", "handlers": ["wsgi"]},
+                "loggers": {
+                    "faker": {"level": "WARNING"},
+                },
+                "disable_existing_loggers": False,
+            },
         },
     }
     return conf