+++ tests/core/test_about.py
@@ -1,2 +1,0 @@
-def test_version(testclient):
-    testclient.get("/about", status=200)