+++ CHANGES.rst
@@ -1,7 +1,10 @@
 Added
 ^^^^^
-
 - Sign in/out events are logged in :issuer:`177`
+
+Fixed
+^^^^^
+- HTMX and JAVASCRIPT configuration settings.
 
 [0.0.49] - 2024-04-08
 ---------------------